pyproject.toml
[tool.poetry]
name = "flo-test"
version = "0.1.0"
description = ""
authors = ["Jules LE BRIS <jls.lebris@gmail.com>"]
readme = "README.md"
package-mode = false


[tool.poetry.group.dev]
optional = true


[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.5"}
pydantic-settings = "^2.6.1"
sqlmodel = "^0.0.22"
alembic = "^1.14.0"
asyncpg = "^0.30.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.10.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
black = "^24.10.0"
isort = "^5.13.2"
flake8 = {extras = ["flake8-pyproject"], version = "^7.1.1"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 119

[tool.flake8]
max-line-length = 119
line-length = 119

[tool.black]
line-length = 119
target-version = ['py311']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''

[tool.coverage.run]
omit = ["src/manage.py"]
source = ["src"]
include = ["src/*"]
branch = true
relative_files = true
command_line = "-m pytest"

[tool.pytest.ini_options]
pythonpath = ["src", "."]
